requirements.txt
numpy
scipy
matplotlib
pyqt5-tools
pyqtgraph
pyaudio